# Add doc-values aggregator for the count(v) aggregation.