## Summary of the changes / Why this improves CrateDB

It improves the performance of the global standard deviation aggregations.
See https://github.com/crate/crate/pull/10048 for more information

Without doc-values aggregators:
```
## Running Query:
   Name: global count on long
   Statement:
     select count(duration) from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [04:14<00:00, 19.67 requests/s]
Runtime (in ms):
    mean:    49.637 ± 0.540
    min/max: 34.369 → 309.033
Percentile:
    50:   43.957 ± 19.476 (stdev)
    95:   77.465
    99.9: 218.348

## Running Query:
   Name: global count on float
   Statement:
     select count("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [04:16<00:00, 19.50 requests/s]
Runtime (in ms):
    mean:    50.090 ± 0.353
    min/max: 41.570 → 225.817
Percentile:
    50:   46.911 ± 12.746 (stdev)
    95:   66.626
    99.9: 219.086
# Running tearDown
```

With doc-values aggregators:
```
# Running setUp
# Running benchmark

## Running Query:
   Name: global count on long
   Statement:
     select count(duration) from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [01:19<00:00, 62.67 requests/s]
Runtime (in ms):
    mean:    14.641 ± 1.044
    min/max: 7.877 → 1172.763
Percentile:
    50:   11.914 ± 37.657 (stdev)
    95:   23.165
    99.9: 226.907

## Running Query:
   Name: global count on float
   Statement:
     select count("adRevenue") from uservisits
   Concurrency: 1
   Iterations: 5000
100%| 5000/5000 [01:10<00:00, 71.00 requests/s]
Runtime (in ms):
    mean:    12.104 ± 0.206
    min/max: 7.777 → 179.364
Percentile:
    50:   10.606 ± 7.430 (stdev)
    95:   19.244
    99.9: 75.315
# Running tearDown
```

## Checklist

 - [ ] Added an entry in `CHANGES.txt` for user facing changes
 - [ ] Updated documentation & `sql_features` table for user facing changes
 - [x] Touched code is covered by tests
 - [x] [CLA](https://crate.io/community/contribute/cla/) is signed
 - [x] This does not contain breaking changes, or if it does:
    - It is released within a major release
    - It is recorded in ``CHANGES.txt``
    - It was marked as deprecated in an earlier release if possible
    - You've thought about the consequences and other components are adapted
      (E.g. AdminUI)