@mfussenegger 2020-08-13
A potential follow up could be to merge the state class with `MutableLong`